Dev

See merge request fajarseniorsys/manajemen-aset!10

+++ app/Http/Controllers/ItemCategoryController.php
@@ -5,7 +5,8 @@ namespace App\Http\Controllers;
 use App\Models\AssetCategory;
 use App\Models\ItemCategory;
 use Illuminate\Http\Request;
-use League\Fractal\Resource\Item;
+use Illuminate\Support\Facades\Log;
+use Illuminate\Support\Facades\Session;
 use Yajra\DataTables\Facades\DataTables;
 
 
@@ -14,6 +15,10 @@ class ItemCategoryController extends Controller
     public function __construct()
     {
         $this->middleware('auth');
+        $this->middleware('permission:kategori-barang-list', ['only' => ['index']]);
+        $this->middleware('permission:kategori-barang-create', ['only' => ['store']]);
+        $this->middleware('permission:kategori-barang-edit', ['only' => ['update']]);
+        $this->middleware('permission:kategori-barang-delete', ['only' => ['destroy']]);
     }
     /**
      * Display a listing of the resource.
@@ -42,13 +47,31 @@ class ItemCategoryController extends Controller
     public function store(Request $request)
     {
         $this->validate($request, [
+            'item_category_name' => 'required|unique:item_categories,item_category_name,NULL,NULL,deleted_at,NULL',
+            'item_category_code' => 'required|unique:item_categories,item_category_code,NULL,NULL,deleted_at,NULL',
             'asset_category_id' => 'required',
-            'item_category_name' => 'required',
+        ],
+        [
+            'item_category_name.unique' => 'Nama barang sudah digunakan',
+            'item_category_code.unique' => 'Kode barang sudah digunakan'
         ]);
-        ItemCategory::create($request->all());
-        return response()->json([
-            'status' => true,
-        ], 200);
+        try {
+            $input = $request->all();
+            $input['item_category_code'] = strtoupper($input['item_category_code']);
+            ItemCategory::create($input);
+            $menu = ItemCategory::get();
+            if (count($menu) > 0) Session::put('categories', $menu);
+            if (count($menu) == 0) Session::put('categories', []);
+            return response()->json([
+                'status' => true,
+            ], 200);
+        } catch (\Throwable $th) {
+            Log::error($th->getMessage());
+            return response()->json([
+                'status' => false,
+                'data' => $th,
+            ], 500);
+        }
     }
 
     /**
@@ -72,13 +95,30 @@ class ItemCategoryController extends Controller
     public function update(Request $request, string $id)
     {
         $this->validate($request, [
+            'item_category_name' => "required|unique:item_categories,item_category_name,$id,item_category_id,deleted_at,NULL",
+            'item_category_code' => "required|unique:item_categories,item_category_code,$id,item_category_id,deleted_at,NULL",
             'asset_category_id' => 'required',
-            'item_category_name' => 'required',
+        ],
+        [
+            'item_category_name.unique' => 'Nama barang sudah digunakan',
+            'item_category_code.unique' => 'Kode barang sudah digunakan'
         ]);
-        ItemCategory::where('item_category_id', $id)->update($request->all());
-        return response()->json([
-            'status' => true,
-        ], 200);
+        $input = $request->all();
+        $input['item_category_code'] = strtoupper($input['item_category_code']);
+        try {
+            ItemCategory::where('item_category_id', $id)->update($input);
+            $menu = ItemCategory::get();
+            if (count($menu) > 0) Session::put('categories', $menu);
+            if (count($menu) == 0) Session::put('categories', []);
+            return response()->json([
+                'status' => true,
+            ], 200);
+        } catch (\Throwable $th) {
+            Log::error($th->getMessage());
+            return response()->json([
+                'status' => false,
+            ], 500);
+        }
     }
 
     /**
@@ -86,9 +126,24 @@ class ItemCategoryController extends Controller
      */
     public function destroy(string $id)
     {
-        ItemCategory::where('item_category_id', $id)->delete();
-        return response()->json([
-            'status' => true,
-        ], 200);
+        try {
+            if ($id == null) {
+                return response()->json([
+                    'status' => false,
+                ], 500);
+            }
+            ItemCategory::where('item_category_id', $id)->delete();
+            $menu = ItemCategory::get();
+            if (count($menu) > 0) Session::put('categories', $menu);
+            if (count($menu) == 0) Session::put('categories', []);
+            return response()->json([
+                'status' => true,
+            ], 200);
+        } catch (\Throwable $th) {
+            Log::error($th->getMessage());
+            return response()->json([
+                'status' => false,
+            ], 500);
+        }
     }
 }
